feature(outputs): add support for machine readable output formats (json,csv)

+++ internal/zillow/client.go
@@ -37,8 +37,8 @@ type PropertyLookupResponse struct {
 	LookupResults []struct {
 		Zpid      int
 		Estimates struct {
-			Zestimate     int
-			RentZestimate int
+			Zestimate     int64
+			RentZestimate int64
 		}
 	}
 }
